refactor: rename from theoretical time to available time
