+++ src/domain/models/available_time/available_time_repository_interface.ts
@@ -1,0 +1,6 @@
+import { AvailableTime } from './available_time';
+
+export interface AvailableTimeRepositoryInterface {
+  create(userId: string, availableTime: number, isoWeek: number): AvailableTime;
+  getAll(): readonly AvailableTime[];
+}
