+++ src/domain/models/summary_report/value_objects/summary_report_theoretical_hour.ts
@@ -1,1 +1,1 @@
-export class SummaryReportTheoreticalHour {}
+export class SummaryReportAvailableHour {}
