+++ src/domain/models/summary_report/value_objects/summary_report_theoretical_rate.ts
@@ -1,1 +1,1 @@
-export class SummaryReportTheoreticalRate {}
+export class SummaryReportAvailableRate {}
