--- src/domain/models/theoretical_time/theoretical_time_repository_interface.ts
@@ -1,10 +1,0 @@
-import { TheoreticalTime } from './theoretical_time';
-
-export interface TheoreticalTimeRepositoryInterface {
-  create(
-    userId: string,
-    theoreticalTime: number,
-    isoWeek: number,
-  ): TheoreticalTime;
-  getAll(): readonly TheoreticalTime[];
-}
