--- src/domain/models/theoretical_time/value_objects/theoretical_time_iso_week.ts
@@ -1,24 +1,0 @@
-export class TheoreticalTimeIsoWeek {
-  private static readonly MAX_ISO_WEEKS_IN_YEAR = 53;
-
-  theoreticalTimeIsoWeek: number;
-  constructor(iosWeek: number) {
-    if (!iosWeek) throw new Error('TheoreticalTimeIsoWeekが存在しません');
-
-    if (isNaN(iosWeek)) {
-      throw new Error('TheoreticalTimeIsoWeekはnumber型でなければなりません');
-    }
-
-    if (TheoreticalTimeIsoWeek.MAX_ISO_WEEKS_IN_YEAR < iosWeek) {
-      throw new Error(
-        `TheoreticalTimeIsoWeekは${TheoreticalTimeIsoWeek.MAX_ISO_WEEKS_IN_YEAR}以下の数値でなければなりません`,
-      );
-    }
-
-    this.theoreticalTimeIsoWeek = iosWeek;
-  }
-
-  toNumber(): number {
-    return this.theoreticalTimeIsoWeek;
-  }
-}
