+++ src/use_case/available_time/calculate/available_time_calculate_input_data.ts
@@ -1,0 +1,15 @@
+import { AvailableTimeInputDataBase } from '../available_time_input_data_base';
+import CalendarEvent = GoogleAppsScript.Calendar.CalendarEvent;
+
+export class AvailableTimeCalculateInputData extends AvailableTimeInputDataBase {
+  getEvents(googleCalendarId: string): CalendarEvent[] {
+    const calendar = CalendarApp.getCalendarById(googleCalendarId);
+    const nextMonday = Moment.moment()
+      .day(1 + 7)
+      .toDate();
+    const nextFriday = Moment.moment()
+      .day(5 + 7)
+      .toDate();
+    return calendar.getEvents(nextMonday, nextFriday);
+  }
+}
