+++ src/use_case/available_time/calculate/available_time_calculate_use_case_interface.ts
@@ -1,0 +1,3 @@
+export interface AvailableTimeCalculateUseCaseInterface {
+  handle(): void;
+}
