+++ src/use_case/available_time/notify/available_time_notify_use_case_interface.ts
@@ -1,0 +1,3 @@
+export interface AvailableTimeNotifyUseCaseInterface {
+  handle(): void;
+}
