--- src/use_case/theoretical_time/calculate/theoretical_time_calculate_use_case_interface.ts
@@ -1,3 +1,0 @@
-export interface TheoreticalTimeCalculateUseCaseInterface {
-  handle(): void;
-}
