--- src/use_case/theoretical_time/notify/theoretical_time_notify_output_data.ts
@@ -1,6 +1,0 @@
-export class TheoreticalTimeNotifyOutputData {
-  getMessage(userId: string, todaysTheoreticalTime: number): string {
-    return `<@${userId}>
-本日の実装可能時間は ${todaysTheoreticalTime / (60 * 60 * 1000)}h です。`;
-  }
-}
