--- src/use_case/theoretical_time/notify/theoretical_time_notify_use_case_interface.ts
@@ -1,3 +1,0 @@
-export interface TheoreticalTimeNotifyUseCaseInterface {
-  handle(): void;
-}
