+++ src/webhook_app/available_time/controllers/calculate/available_time_calculate_controller.ts
@@ -1,0 +1,11 @@
+import { AvailableTimeCalculateUseCaseInterface } from '../../../../use_case/available_time/calculate/available_time_calculate_use_case_interface';
+
+export class AvailableTimeCalculateController {
+  constructor(
+    private readonly availableTimeCalculateUseCase: AvailableTimeCalculateUseCaseInterface,
+  ) {}
+
+  calculate(): void {
+    this.availableTimeCalculateUseCase.handle();
+  }
+}
