--- src/webhook_app/theoretical_time/controllers/calculate/theoretical_time_calculate_controller.ts
@@ -1,11 +1,0 @@
-import { TheoreticalTimeCalculateUseCaseInterface } from '../../../../use_case/theoretical_time/calculate/theoretical_time_calculate_use_case_interface';
-
-export class TheoreticalTimeCalculateController {
-  constructor(
-    private readonly theoreticalTimeCalculateUseCase: TheoreticalTimeCalculateUseCaseInterface,
-  ) {}
-
-  calculate(): void {
-    this.theoreticalTimeCalculateUseCase.handle();
-  }
-}
