--- src/webhook_app/theoretical_time/controllers/notify/theoretical_time_notify_controller.ts
@@ -1,11 +1,0 @@
-import { TheoreticalTimeNotifyUseCaseInterface } from '../../../../use_case/theoretical_time/notify/theoretical_time_notify_use_case_interface';
-
-export class TheoreticalTimeNotifyController {
-  constructor(
-    private readonly theoreticalTimeNotifyUseCase: TheoreticalTimeNotifyUseCaseInterface,
-  ) {}
-
-  notify(): void {
-    this.theoreticalTimeNotifyUseCase.handle();
-  }
-}
